refactor: rename the taskMenuStatus file
